Move sop function to shared.js
