+++ utils/shared.js
@@ -691,6 +691,21 @@ const swapObjKeysAndValues = (object) => {
     return newObject;
 }
 
+const sop = async (req, res) => {
+    setHeadersDomainRestricted(req, res);
+
+    if(req.method === 'OPTIONS') return res.status(200).json({code: 200});
+
+    if(req.method !== 'POST') {
+        return res.status(405).json(getResponseJSON('Only POST requests are accepted!', 405));
+    }
+
+    const { retrieveParticipantsDataDestruction } = require(`./firestore`);
+    await retrieveParticipantsDataDestruction()
+
+    return res.status(200).json({message: 'Success!', code: 200})
+}
+
 module.exports = {
     getResponseJSON,
     setHeaders,
@@ -722,5 +737,6 @@ module.exports = {
     cleanSurveyData,
     refusalWithdrawalConcepts,
     convertSiteLoginToNumber,
-    swapObjKeysAndValues,
+    swapObjKeysAndValues,,
+    sop
 }
